+++ src/Tomix.Provider.Tom/TomPropertyApplier.cs
@@ -110,6 +110,9 @@ internal static class TomPropertyApplier
             case CalculationItem item when isMainExpression:
                 item.Expression = edit.Value;
                 break;
+            case Function function when isMainExpression:
+                function.Expression = edit.Value;
+                break;
             case Partition { Source: CalculatedPartitionSource source } when isMainExpression:
                 source.Expression = edit.Value;
                 break;
@@ -491,6 +494,18 @@ internal static class TomPropertyApplier
             case "expression":
                 expression.Expression = value;
                 break;
+            case "kind":
+                expression.Kind = ParseEnum<ExpressionKind>(value, displayName);
+                break;
+            case "remoteparametername":
+                expression.RemoteParameterName = value;
+                break;
+            case "lineagetag":
+                expression.LineageTag = value;
+                break;
+            case "sourcelineagetag":
+                expression.SourceLineageTag = value;
+                break;
             default:
                 throw new NotSupportedException($"Setting '{displayName}' is not supported for expressions.");
         }
@@ -511,6 +526,12 @@ internal static class TomPropertyApplier
                 break;
             case "ishidden":
                 function.IsHidden = ParseBool(value, displayName);
+                break;
+            case "lineagetag":
+                function.LineageTag = value;
+                break;
+            case "sourcelineagetag":
+                function.SourceLineageTag = value;
                 break;
             default:
                 throw new NotSupportedException($"Setting '{displayName}' is not supported for functions.");